- Renamed NexusAPI to NexusApiClient for to match naming of MegaApi. - Moved NexusApiClient and all of its Dtos into their own package. - Moved util methods into a new util class. - NexusApi no longer creates new HttpClients for each request and uses one internal HttpClient instead. - Added rudimentary tracking of rate limits.
