+++ Wabbajack/Installer.cs
@@ -14,6 +14,7 @@ using Compression.BSA;
 using K4os.Compression.LZ4.Streams;
 using VFS;
 using Wabbajack.Common;
+using Wabbajack.NexusApi;
 using Directory = Alphaleonis.Win32.Filesystem.Directory;
 using File = Alphaleonis.Win32.Filesystem.File;
 using FileInfo = Alphaleonis.Win32.Filesystem.FileInfo;
@@ -24,6 +25,8 @@ namespace Wabbajack
     public class Installer
     {
         private string _downloadsFolder;
+
+        private NexusApiClient _nexusApiClient;
 
         public Installer(ModList mod_list, string output_folder, Action<string> log_fn)
         {
@@ -46,7 +49,6 @@ namespace Wabbajack
         public Action<string> Log_Fn { get; }
         public Dictionary<string, string> HashedArchives { get; private set; }
 
-        public string NexusAPIKey { get; set; }
         public bool IgnoreMissingFiles { get; internal set; }
         public string GameFolder { get; set; }
 
@@ -169,7 +171,7 @@ namespace Wabbajack
 
             mods.PMap(mod =>
             {
-                var er = NexusAPI.EndorseMod(mod, NexusAPIKey);
+                var er = _nexusApiClient.EndorseMod(mod);
                 Utils.Log($"Endorsed {mod.GameName} - {mod.ModID} - Result: {er.message}");
             });
             Info("Done! You may now exit the application!");
@@ -409,14 +411,19 @@ namespace Wabbajack
             Info("Getting Nexus API Key, if a browser appears, please accept");
             if (ModList.Archives.OfType<NexusMod>().Any())
             {
-                NexusAPIKey = NexusAPI.GetNexusAPIKey();
+                _nexusApiClient = new NexusApiClient();
 
-                var user_status = NexusAPI.GetUserStatus(NexusAPIKey);
-
-                if (!user_status.is_premium)
+                if (!_nexusApiClient.IsAuthenticated)
                 {
-                    Info(
-                        $"Automated installs with Wabbajack requires a premium nexus account. {user_status.name} is not a premium account");
+                    Error(
+                        $"Authenticating for the Nexus failed. A nexus account is required to automatically download mods.");
+                    return;
+                }
+
+                if (!_nexusApiClient.IsPremium)
+                {
+                    Error(
+                        $"Automated installs with Wabbajack requires a premium nexus account. {_nexusApiClient.Username} is not a premium account.");
                     return;
                 }
             }
@@ -449,7 +456,7 @@ namespace Wabbajack
                         string url;
                         try
                         {
-                            url = NexusAPI.GetNexusDownloadLink(a, NexusAPIKey, !download);
+                            url = _nexusApiClient.GetNexusDownloadLink(a, !download);
                             if (!download) return true;
                         }
                         catch (Exception ex)
