+++ Wabbajack/Wabbajack.csproj
@@ -182,6 +182,8 @@
     </ApplicationDefinition>
     <Compile Include="Data.cs" />
     <Compile Include="LambdaCommand.cs" />
+    <Compile Include="NexusApi\Dtos.cs" />
+    <Compile Include="NexusApi\NexusApiUtils.cs" />
     <Compile Include="ReportBuilder.cs" />
     <Compile Include="Themes\LeftMarginMultiplierConverter.cs" />
     <Compile Include="Themes\TreeViewItemExtensions.cs" />
@@ -209,7 +211,7 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="NexusAPI.cs" />
+    <Compile Include="NexusApi\NexusApi.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
